fix: resolve clippy lint_groups_priority, trivial casts, and elided lifetimes
- Set lint group priorities in Cargo.toml to allow individual lint overrides
- Remove unused extern crate proc_macro in acton-macro
- Fix unnecessary qualifications (crate::common:: prefixes)
- Replace trivial 'as' casts with explicit type annotations
- Add explicit lifetime parameter to Formatter in Display impl
- Remove trivial u64 cast for agent_shutdown_timeout_ms

+++ acton-reactive/src/actor/managed_agent/started.rs
@@ -302,7 +302,7 @@ impl<Agent: Default + Send + Debug + 'static> ManagedAgent<Started, Agent> {
         use std::time::Duration;
         use tokio::time::timeout as tokio_timeout;
 
-        let timeout_ms = CONFIG.timeouts.agent_shutdown_timeout_ms as u64;
+        let timeout_ms = CONFIG.timeouts.agent_shutdown_timeout_ms;
 
         let stop_futures: Vec<_> = self
             .handle
